@@ -1,11 +1,4 @@
-[aliases]
-dists = clean --all sdist bdist_wheel
-
-[bdist_wheel]
-universal = 1
-
 [metadata]
-license_file = LICENSE
 name = jaraco.mongodb
 author = Jason R. Coombs
 author_email = jaraco@jaraco.com
@@ -16,49 +9,69 @@ classifiers =
 	Development Status :: 5 - Production/Stable
 	Intended Audience :: Developers
 	License :: OSI Approved :: MIT License
-	Programming Language :: Python :: 2.7
 	Programming Language :: Python :: 3
+	Programming Language :: Python :: 3 :: Only
 	Framework :: Pytest
 
 
 [options]
-packages = find:
-namespace_packages = jaraco
+packages = find_namespace:
 include_package_data = true
-python_requires = >=2.7,!=3.0,!=3.1,!=3.2
+python_requires = >=3.8
 install_requires =
 	pymongo>=3.5
-	six
 	python-dateutil
-	jaraco.services
+	jaraco.services>=2
 	portend
-	jaraco.itertools>=1.8
-	jaraco.functools
-	jaraco.ui>=1.6
-	jaraco.context
+	jaraco.itertools>=2
+	jaraco.functools>=2
+	jaraco.ui>=2
+	jaraco.context>=2
 	more_itertools
-	jaraco.logging
-	jaraco.timing
+	jaraco.logging>=2
+	tempora
 	pytimeparse
-	jaraco.collections>=1.5
-setup_requires = setuptools_scm >= 1.15.0
+	jaraco.collections>=2
+	importlib_metadata; python_version < "3.8"
+	autocommand
+	cachetools
+
+[options.packages.find]
+exclude =
+	build*
+	dist*
+	docs*
+	tests*
 
 [options.extras_require]
 testing =
 	# upstream
-	pytest >= 3.5, !=3.7.3
-	pytest-sugar >= 0.9.1
-	collective.checkdocs
-	pytest-flake8
+	pytest >= 6
+	pytest-checkdocs >= 2.4
+	pytest-black >= 0.3.7; \
+		# workaround for jaraco/skeleton#22
+		python_implementation != "PyPy"
+	pytest-cov
+	pytest-mypy >= 0.9.1; \
+		# workaround for jaraco/skeleton#22
+		python_implementation != "PyPy"
+	pytest-enabler >= 2.2
+	pytest-ruff
 
 	# local
 	cherrypy
+	types-python-dateutil
+	types-cachetools
 
 docs =
 	# upstream
-	sphinx
-	jaraco.packaging >= 3.2
+	sphinx >= 3.5
+	# workaround for sphinx/sphinx-doc#11662
+	sphinx < 7.2.5
+	jaraco.packaging >= 9.3
 	rst.linker >= 1.9
+	furo
+	sphinx-lint
 
 	# local
 	cherrypy
