@@ -1,32 +1,47 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
 extensions = [
-	'sphinx.ext.autodoc',
-	'jaraco.packaging.sphinx',
-	'rst.linker',
+    'sphinx.ext.autodoc',
+    'jaraco.packaging.sphinx',
 ]
 
-master_doc = 'index'
+master_doc = "index"
+html_theme = "furo"
 
+# Link dates and other references in the changelog
+extensions += ['rst.linker']
 link_files = {
-	'../CHANGES.rst': dict(
-		using=dict(
-			GH='https://github.com',
-		),
-		replace=[
-			dict(
-				pattern=r'(Issue #|\B#)(?P<issue>\d+)',
-				url='{package_url}/issues/{issue}',
-			),
-			dict(
-				pattern=r'^(?m)((?P<scm_version>v?\d+(\.\d+){1,2}))\n[-=]+\n',
-				with_scm='{text}\n{rev[timestamp]:%d %b %Y}\n',
-			),
-			dict(
-				pattern=r'PEP[- ](?P<pep_number>\d+)',
-				url='https://www.python.org/dev/peps/pep-{pep_number:0>4}/',
-			),
-		],
-	),
+    '../NEWS.rst': dict(
+        using=dict(GH='https://github.com'),
+        replace=[
+            dict(
+                pattern=r'(Issue #|\B#)(?P<issue>\d+)',
+                url='{package_url}/issues/{issue}',
+            ),
+            dict(
+                pattern=r'(?m:^((?P<scm_version>v?\d+(\.\d+){1,2}))\n[-=]+\n)',
+                with_scm='{text}\n{rev[timestamp]:%d %b %Y}\n',
+            ),
+            dict(
+                pattern=r'PEP[- ](?P<pep_number>\d+)',
+                url='https://peps.python.org/pep-{pep_number:0>4}/',
+            ),
+        ],
+    )
 }
+
+# Be strict about any broken references
+nitpicky = True
+
+# Include Python intersphinx mapping to prevent failures
+# jaraco/skeleton#51
+extensions += ['sphinx.ext.intersphinx']
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+}
+
+# Preserve authored syntax for defaults
+autodoc_preserve_defaults = True
+
+intersphinx_mapping.update(
+    cherrypy=('https://docs.cherrypy.dev/en/latest/', None),
+    pymongo=('https://pymongo.readthedocs.io/en/stable/', None),
+)
